requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
requests
fastapi
boto3
awscli
pushbullet.py
papermill
nbconvert
weasyprint
pdfkit
